Adding additional Integration Tests for the 2.0 client.

+++ src/Postmark/Model/PostmarkOutboundOverviewStats.cs
@@ -6,9 +6,9 @@ namespace PostmarkDotNet.Model
         public int Sent { get; set; }
         public int Bounced { get; set; }
         public int SMTPApiErrors { get; set; }
-        public int BounceRate { get; set; }
+        public double BounceRate { get; set; }
         public int SpamComplaints { get; set; }
-        public int SpamComplaintsRate { get; set; }
+        public double SpamComplaintsRate { get; set; }
         public int Opens { get; set; }
         public int UniqueOpens { get; set; }
         public int Tracked { get; set; }
